/update make all reservation 5ryu messages

+++ src/services/reservation.service.js
@@ -6,30 +6,21 @@ class ReservationService {
   }
 
   // 예약 생성 서비스 함수
-  async createReservationService(userId, petsitterId, reservationDate, reservationStatus) {
+  async createReservationService(userId, petsitterId, reservationDate) {
     // 펫시터가 존재하는지 확인
     const petsitter = await this.reservationRepository.findPetsitterById(petsitterId);
     if (!petsitter) {
       throw new HttpError.NotFound('해당 펫시터는 존재하지 않습니다.');
     }
-    // 이미 예약된 날짜인지 확인 중복날짜면 예약이 불가합니당
-    const isAlreadyReserved = await this.reservationRepository.AlreadyReserved(
+    // 예약상태가 ACCEPT면 예약생성이 불가합니다.
+    const alreadyAcceptStatus = await this.reservationRepository.acceptStatus(
       petsitterId,
       reservationDate
     );
-    if (isAlreadyReserved) {
-      throw new HttpError.Conflict('해당 날짜는 예약이 불가한 날짜입니다.');
+    if (alreadyAcceptStatus) {
+      throw new HttpError.Conflict('상태가 ACCEPT인 예약이 이미 존재합니다.');
     }
-    // 예약상태가 존재하면 예약이 불가합니다.
-    const unacceptableStatus = await this.reservationRepository.UnacceptableStatus(
-      petsitterId,
-      reservationDate,
-      reservationStatus
-    );
-    if (unacceptableStatus) {
-      console.log(unacceptableStatus);
-      throw new HttpError.Conflict('이미예약있어요.');
-    }
+
     const reservationData = {
       userId,
       petsitterId,
@@ -38,23 +29,68 @@ class ReservationService {
     return await this.reservationRepository.createReservation(reservationData);
   }
 
-  // 예약 목록 조회 서비스 함수
+  // 나의 예약 목록 조회 서비스 함수
   async getReservationService(userId, sort) {
+    // 나의 예약이 없으면 조회가 안됨 남의거 보이면 안됨
+    const noReservation = await this.reservationRepository.getReservation(userId, sort);
+    if (noReservation.length === 0) {
+      throw new HttpError.NotFound('예약내역이 존재하지 않습니다.');
+    }
     return await this.reservationRepository.getReservation(userId, sort);
   }
 
   // 예약 상세 조회 서비스 함수
   async getReservationByIdService(reservationId, userId) {
+    const notMyReservation = await this.reservationRepository.getReservationById(
+      reservationId,
+      userId
+    );
+    if (!notMyReservation || notMyReservation.userId !== userId) {
+      throw new HttpError.NotFound('본인의 예약 내역만 조회가 가능합니다.');
+    }
     return await this.reservationRepository.getReservationById(Number(reservationId), userId);
   }
 
   // 예약 수정 서비스 함수
   async updateReservationService(reservationId, userId, data) {
-    return await this.reservationRepository.updateReservation(Number(reservationId), userId, data);
+    // reservationStatus 필드를 제거합니다.
+    // 변경된 부분: reservationStatus 변수를 할당만 하고 사용하지 않기 때문에 아래메시지
+    //eslint-disable-next-line no-unused-vars
+    const { reservationStatus, ...updateData } = data;
+    //자신의 예약 내역이 아니면 수정이 불가합니다.
+    const reservation = await this.reservationRepository.getReservationById(reservationId, userId);
+    if (!reservation || reservation.userId !== userId) {
+      throw new HttpError.NotFound('본인의 예약 내역만 수정이 가능합니다.');
+    }
+    // 예약 날짜가 제공되지 않은 경우 오류를 던집니다.
+    if (!updateData.reservationDate || isNaN(new Date(updateData.reservationDate).getTime())) {
+      throw new HttpError.BadRequest('유효한 예약 날짜가 제공되지 않았습니다.');
+    }
+    // 예약상태가 AWAIT가 아니면 수정이 불가합니다.
+    const cantUpdateStatus = await this.reservationRepository.awaitStatus(reservationId, userId);
+    if (!cantUpdateStatus) {
+      throw new HttpError.Conflict('예약상태가 AWAIT가 아니면 수정이 불가합니다.');
+    }
+
+    return await this.reservationRepository.updateReservation(
+      Number(reservationId),
+      userId,
+      updateData
+    );
   }
 
   // 예약 삭제 서비스 함수
   async deleteReservationService(reservationId, userId) {
+    //자신의 예약 내역이 아니면 삭제가 불가합니다.
+    const reservation = await this.reservationRepository.getReservationById(reservationId, userId);
+    if (!reservation || reservation.userId !== userId) {
+      throw new HttpError.NotFound('본인의 예약 내역만 삭제 가능합니다.');
+    }
+    // 예약상태가 AWAIT가 아니면 삭제가 불가합니다.
+    const cantUpdateStatus = await this.reservationRepository.awaitStatus(reservationId, userId);
+    if (!cantUpdateStatus) {
+      throw new HttpError.Conflict('예약상태가 AWAIT가 아니면 삭제가 불가합니다.');
+    }
     return await this.reservationRepository.deleteReservation(Number(reservationId), userId);
   }
 }
